pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai>=0.6.3",
    "google-search-results==2.4.2",
    "litellm==1.72.6",
    "pandas>=2.3.0",
    "pandas-stubs==2.2.3.250527",
    "phonenumbers==9.0.7",
    "pyap2==0.1.14",
]
